requirements.txt
# Core Data Science Libraries
pandas>=1.3.0
numpy>=1.21.0

# Machine Learning
scikit-learn>=1.0.0

# Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Jupyter Notebook Support
jupyter>=1.0.0
ipython>=7.0.0

# Document Conversion (optional)
python-docx>=0.8.11

# Utility
tqdm>=4.62.0
